Revamp test framework to be in-process and cleaner
