+++ promflux/testdata/prom-data/01D48W68D1WMK28VP4P3S6K5FN/tombstones
